file tools refactored
